Moved testing classes to test package
