+++ .gitignore
@@ -111,4 +111,3 @@ local.properties
 
 # virtual machine crash logs, see http://www.java.com/en/download/help/error_hotspot.xml
 hs_err_pid*
-.idea/workspace.xml
