+++ gui.iml
@@ -97,8 +97,15 @@
     <orderEntry type="library" scope="TEST" name="Maven: com.sun.jersey:jersey-client:1.19" level="project" />
     <orderEntry type="library" scope="TEST" name="Maven: com.sun.jersey:jersey-core:1.19" level="project" />
     <orderEntry type="library" scope="TEST" name="Maven: org.neo4j.app:neo4j-server:test-jar:tests:3.1.0" level="project" />
-    <orderEntry type="library" scope="TEST" name="Maven: org.neo4j.driver:neo4j-java-driver:1.1.0" level="project" />
+    <orderEntry type="library" name="Maven: org.slf4j:slf4j-log4j12:1.7.21" level="project" />
+    <orderEntry type="library" name="Maven: org.slf4j:slf4j-api:1.7.21" level="project" />
+    <orderEntry type="library" name="Maven: log4j:log4j:1.2.17" level="project" />
+    <orderEntry type="library" name="Maven: org.neo4j.driver:neo4j-java-driver:1.1.1" level="project" />
     <orderEntry type="library" scope="TEST" name="Maven: junit:junit:4.12" level="project" />
     <orderEntry type="library" scope="TEST" name="Maven: org.hamcrest:hamcrest-core:1.3" level="project" />
+    <orderEntry type="library" name="Maven: org.jetbrains:annotations:15.0" level="project" />
+    <orderEntry type="library" name="Maven: org.junit.jupiter:junit-jupiter-api:5.0.0-M3" level="project" />
+    <orderEntry type="library" name="Maven: org.opentest4j:opentest4j:1.0.0-M1" level="project" />
+    <orderEntry type="library" name="Maven: org.junit.platform:junit-platform-commons:1.0.0-M3" level="project" />
   </component>
 </module>
